third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-step/debugger-step-out-expected.txt

Tests "step out" functionality in debugger.

Set timer for test function.
Script execution paused.
Call stack:
    0) d (debugger-step-out.js:17)
    1) testFunction (debugger-step-out.js:22)
Stepping out...
Script execution resumed.
Script execution paused.
Call stack:
    0) testFunction (debugger-step-out.js:23)
Script execution resumed.
